wasteskipsblantyre: Locations of Public Waste Skips in Blantyre, Malawi

An R data package containing the locations of public waste skips in Blantyre, Malawi, updated in 2023, with the number of skips, vehicle access and the fill-up times reported at each location, and the points of interest for their collection (depot, dump site, compost site and gas stations).

Author

Maintainer: Lars Schöbitz lschoebitz@ethz.ch (ORCID) [contributor]

Authors:

Mabvuto Yesaya yesayamabvuto@gmail.com (ORCID)

Limbani Msuku limbani_msuku@yahoo.com (ORCID)

Elizabeth Tilley tilleye@ethz.ch (ORCID)

Sebastian Camilo Loos seba.loos@hotmail.com (ORCID)

Nicolas Seemann-Ricard nseemann@ethz.ch (ORCID)

Other contributors:

Mian Zhong mzhong@ethz.ch (ORCID) [contributor]